Reintroduce Log4RS as the logging backend for now.
Im still not happy with the look and feel that rocket provides by default
(getting debug or trace output for the projec would mean getting the same
for hyper, which will generate an absurd amount of stuff, some of which
follows a seperate line format, and its just generally a mess. I'm not
too sure what theyre using to handle logging, and itll likely get sorted
out at some point (there is an issue still open, at least) but for now
this gives us something most people will probably be used to, and we can
control the level on a per-project or per-module basis, which is nice.

By bootstrapping it in before anything else in the main function we make
sure that our stuff gets used by rocket as welll.